# ProductTraceability_2026/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Agricultural Traceability System API
  version: 1.0.0
  description: API specification for the agricultural product traceability system.
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: http://localhost:8888
paths:
...
  /api/batch/list:
    get:
      summary: "[Full] Batch list"
      tags: [Batch]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: pageNum, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: integer } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: productType, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Paginated batch list
  /api/batch/detail:
    get:
      summary: "[Full] Batch detail"
      tags: [Batch]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: batchCode, in: query, schema: { type: string } }
        - { name: id, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Batch detail
  /api/batch/status-count:
    get:
      summary: "[Full] Batch status count"
      tags: [Batch]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Count by status
  /api/batch/create:
    post:
      summary: "[Full] Create batch - auto-generates batchCode and qrCodeUrl"
      description: |
        Create a new trace batch.
        Auto-fills: batchCode, productName/unit defaults, productionDate, qrCodeUrl.
      tags: [Batch]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productType, enterpriseId]
              properties:
                productType: { type: integer, description: "1-chicken, 2-tomato" }
                productName: { type: string, description: "Auto-filled if empty" }
                breed: { type: string }
                quantity: { type: number }
                unit: { type: string, description: "Auto-filled if empty" }
                originLocation: { type: string }
                latitude: { type: number }
                longitude: { type: number }
                manager: { type: string }
                plantArea: { type: string }
                greenhouseNo: { type: string }
                seedSource: { type: string }
                enterpriseId: { type: integer }
                productionDate: { type: string, format: date, description: "Defaults to today" }
                expectedHarvestDate: { type: string, format: date }
      responses:
        '200':
          description: "Returns id, batchCode, qrCodeUrl"
  /api/batch/quarantine-check/{batchCode}:
    get:
      summary: "[Full] Check quarantine status"
      tags: [Batch]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: batchCode, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Quarantine check result
  /api/batch/receive:
    post:
      summary: "[Full] Receive batch (processor)"
      tags: [Batch]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [batchCode]
              properties:
                batchCode: { type: string }
                batchId: { type: integer }
      responses:
        '200':
          description: Success

  # ==========================================
  # Growth Records
  # ==========================================
...
  /api/dashboard/farmer-stats:
    get:
      summary: "[Full] Farmer enterprise homepage stats"
      tags: [Dashboard]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: enterpriseId, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          description: activeBatchCount, totalLivestock, totalOnChain, warningCount, dailyTrend
  /api/dashboard/processor-stats:
    get:
      summary: "[Full] Processor enterprise homepage stats"
      description: |
        Returns KPI cards for type2 (processor) enterprise dashboard:
        todayReceived (batch count), todayProcessed (kg), pendingSaleInventory (kg),
        totalOnChain, processingTrend (7-day line), conversionRate (pie: input vs output).
      tags: [Dashboard]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: enterpriseId, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          description: Processor dashboard stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  data:
                    type: object
                    properties:
                      todayReceived: { type: integer, description: "Today inbound batch count" }
                      todayProcessed: { type: number, description: "Today output quantity (kg)" }
                      pendingSaleInventory: { type: number, description: "Processed but unsold inventory (kg)" }
                      totalOnChain: { type: integer }
                      processingTrend:
                        type: array
                        items:
                          type: object
                          properties:
                            date: { type: string }
                            amount: { type: number }
                      conversionRate:
                        type: array
                        items:
                          type: object
                          properties:
                            name: { type: string }
                            value: { type: number }
  /api/dashboard/analysis-stats:
    get:
      summary: "[Full] Data analysis statistics page"
      description: |
        Returns KPI cards (activeBatchCount, totalLivestock, passRate),
        inputTrend (30-day input usage), statusDistribution (pie),
        chainActivity (30-day on-chain count bar).
      tags: [Dashboard]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: enterpriseId, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          description: Analysis stats aggregation
  /api/dashboard/admin-stats:
    get:
      summary: "✅ Admin dashboard statistics"
      tags: [Dashboard]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: role, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: type1Count, type2Count, type3Count, totalBatches, batchStatusCounts, monthlyBatches

  # ==========================================
  # Enterprise List Management (Admin)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...